Added testCase for all Services
