+++ RestEasy Example/src/main/java/com/baeldung/server/MovieCrudService.java
@@ -18,17 +18,20 @@ public class MovieCrudService {
 
     private Map<String,Movie> inventory = new HashMap<String, Movie>();
 
+
     @GET
     @Path("/getinfo")
     @Produces({MediaType.APPLICATION_JSON,MediaType.APPLICATION_XML})
     public Movie movieByImdbID(@QueryParam("imdbID") String imdbID){
 
-        System.out.println("*** Calling  getinfo ***");
+        System.out.println("*** Calling  getinfo for a given ImdbID***");
 
-        Movie movie=new Movie();
-        movie.setImdbID(imdbID);
-        return movie;
+        if(inventory.containsKey(imdbID)){
+            return inventory.get(imdbID);
+        }else return null;
+
     }
+
 
     @POST
     @Path("/addmovie")
@@ -41,6 +44,7 @@ public class MovieCrudService {
             return Response.status(Response.Status.NOT_MODIFIED)
                     .entity("Movie is Already in the database.").build();
         }
+
         inventory.put(movie.getImdbID(),movie);
 
         return Response.status(Response.Status.CREATED).build();
@@ -54,7 +58,7 @@ public class MovieCrudService {
 
         System.out.println("*** Calling  updateMovie ***");
 
-        if (null!=inventory.get(movie.getImdbID())){
+        if (null==inventory.get(movie.getImdbID())){
             return Response.status(Response.Status.NOT_MODIFIED)
                     .entity("Movie is not in the database.\nUnable to Update").build();
         }
@@ -78,6 +82,7 @@ public class MovieCrudService {
         inventory.remove(imdbID);
         return Response.status(Response.Status.OK).build();
     }
+
 
     @GET
     @Path("/listmovies")
